User-space C library in Rust and assembler
Provide a super-subset of the userspace C library, implemented
mostly in Rust with some assembler required.

Not all of the functions from the ISO C library are provided.
Of those provided, some are minimal and incomplete and others
have different signatures than one might expect from POSIX or
ISO C.  Some functions defined in neither POSIX nor ISO C are
provided in lieu of standard functions.  E.g., we eschew
`strcpy` in favor of `strlcpy`.

Most of the code is written in pure Rust though some is
necessarily written in assembler: the obvious examples are the
system call stubs to request services from the kernel.

Suprisingly, however, the `printf` function is written in
assembly since Rust can declare but not define variadic
functions.  Thus, while Rust fortunately exposes intrinsics to
extract arguments from `va_list` structures and we provide a
pure Rust version of a `vprintf` subset, the `printf` function
is a stub that constructs a `va_list` as defined by the ABI and
passes it to `vprintf`.  We provide this assembly language
version instead of the trivial C stub in order to illustrate how
the variable argument list functionality is implemented; at
least partially.

Some of the provided functions probably have bugs or could be
improved in various ways, be it performance, style, or idiom.
For example, the K&R-derived implementation of malloc and free
is essentially transliterated from the C code and makes little
use of the power of Rust.  Further, that allocator is known to
be generally poorly suited to modern applications.  User beware;
there be dragons here.

Signed-off-by: Dan Cross <cross@gajendra.net>